// src/types/FlavorizationRuleDTO.ts
export type FlavorizationRuleDTO = {
  rowIndex: number;

  disabled: string;
  id: string;
  flavorizationLevel: string;
  match: string;
  flags: string;
  partOfSpeech: string;
  genesis: string;

  replacement1: string;
  replacement2: string;
  replacement3: string;
  replacement4: string;
  replacement5: string;
};